💄 style: opencv loading component 분리
