+++ src/screens/Main/Image2VideoScreen.tsx
@@ -5,7 +5,7 @@ import Video from 'react-native-video';
 import MaterialIcons from 'react-native-vector-icons/MaterialCommunityIcons';
 import { heightSelector, uriSelector, widthSelector } from '../../recoil/selector';
 import { useRecoilValue } from 'recoil';
-import GifImage from '@lowkeychat/react-native-gif';
+import OpenCV from '../../components/Main/Loading/OpenCV';
 
 const Image2VideoScreen = () => {
   const video = require('../../assets/video/output2.mp4');
@@ -21,29 +21,10 @@ const Image2VideoScreen = () => {
   const resizeHeight = (screenWidth*originalHeight*ratio) / originalWidth;
 
   const [isLoading, setIsLoading] = useState(true);
-  const loadingGif = 'file:///Users/heehminh/src/ArtFront-ReactNative/src/assets/gif/landscape0.gif';
   
-  // TODO 서버연결 로딩 컴포넌트
-  // const filePath = storeGifRNFS({ response })
-  // const loadingGif = `file://${filePath}`;
-
-  useEffect(()=>{
-    const timer = setTimeout(()=>{
-      setIsLoading(false);
-    }, 20000);
-
-    return () => clearTimeout(timer);
-  }, []);
-
   if (isLoading) {
     return (
-      <View style={{backgroundColor:'white'}}>
-      <Header nextPage='MainScreen' />
-      <View style={{"backgroundColor": 'white', width: "100%", height: "100%", display: 'flex', marginTop:150,  alignItems: 'center'}}>
-        <GifImage source={{uri: loadingGif}} style={{zIndex:1, width: 400, height: 400 }} resizeMode={'contain'} />
-        <Text style={{marginTop: 30}}>그림의 생명을 불어넣는중..</Text>
-      </View>
-      </View>
+      <OpenCV setIsLoading={setIsLoading} />
     )
   }
